ml/service/pyproject.toml
[project]
name = "aether-ml"
version = "0.1.0"
description = "AETHER AI model routing and inference service"
requires-python = ">=3.12"
dependencies = [
  "accelerate>=1.10.0",
  "datasets>=4.0.0",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "joblib>=1.5.1",
  "peft>=0.17.0",
  "pydantic-settings>=2.10.1",
  "safetensors>=0.6.2",
  "scikit-learn>=1.7.1",
  "sentence-transformers>=5.1.0",
  "torch>=2.8.0",
  "transformers>=4.55.0",
  "uvicorn[standard]>=0.35.0"
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.12.10"
]
